refactor: move cli out of experimental to rllm/cli
Move rllm/experimental/cli/ to rllm/cli/ and update all imports across
source files, tests, and pyproject.toml entry point. The CLI is stable
and no longer belongs under experimental.
