+++ pyproject.toml
@@ -44,7 +44,7 @@ dependencies = [
 ]
 
 [project.scripts]
-rllm = "rllm.experimental.cli.main:cli"
+rllm = "rllm.cli.main:cli"
 
 [project.optional-dependencies]
 
@@ -220,7 +220,7 @@ ignore = [
 "examples/sdk/solver_judge/solver_judge_flow_decorator.py" = ["E501"]
 "examples/sdk/solver_judge/solver_judge_flow_session.py" = ["E501"]
 "examples/solver_judge/solver_judge_flow.py" = ["E501"]
-"rllm/experimental/cli/main.py" = ["E501"]
+"rllm/cli/main.py" = ["E501"]
 "rllm/experimental/rllm_telemetry/examples/experiment_comparison.py" = ["E501"]
 "rllm/experimental/test_examples/opsd/math_opsd_workflow.py" = ["E501"]
 "rllm/experimental/agents/tools/file_editor_tool.py" = ["E501"]
