+++ rllm/cli/eval.py
@@ -17,7 +17,7 @@ from rich.status import Status
 from rich.table import Table
 from rich.theme import Theme
 
-from rllm.experimental.cli._pull import load_dataset_catalog, pull_dataset
+from rllm.cli._pull import load_dataset_catalog, pull_dataset
 
 theme = Theme({"label": "dim", "success": "bold green", "error": "bold red", "val": "bold", "key": "yellow"})
 console = Console(theme=theme)
@@ -97,7 +97,7 @@ def _run_eval(
 
         # Explicit Harbor prefix: "harbor:<name>" resolves from Harbor registry
         if catalog_entry is None and benchmark.startswith("harbor:"):
-            from rllm.experimental.cli._pull import resolve_harbor_catalog_entry
+            from rllm.cli._pull import resolve_harbor_catalog_entry
 
             harbor_name = benchmark.removeprefix("harbor:")
             with Status(f"[dim]Looking up '{harbor_name}' in Harbor registry...[/]", console=console):
@@ -215,7 +215,7 @@ def _run_eval(
     # Resolve agent description
     agent_desc = ""
     if ":" not in agent_name:
-        from rllm.experimental.cli._pull import load_agent_catalog
+        from rllm.cli._pull import load_agent_catalog
 
         agent_catalog = load_agent_catalog()
         agent_entry = agent_catalog.get("agents", {}).get(agent_name, {})
