+++ rllm/cli/model_cmd.py
@@ -13,7 +13,7 @@ from rich.panel import Panel
 from rich.prompt import Confirm, Prompt
 from rich.table import Table
 
-from rllm.experimental.cli._ui import (
+from rllm.cli._ui import (
     _mask_key,
     _prompt_base_url,
     _select_model,
