+++ rllm/cli/templates/crewai_agent.py.tpl
