+++ rllm/cli/templates/custom_evaluator.py.tpl
