+++ rllm/cli/templates/init_py.tpl
