+++ rllm/cli/templates/minimal_agent.py.tpl
