+++ rllm/cli/templates/openai_agents_agent.py.tpl
