+++ rllm/cli/templates/react_agent.py.tpl
