+++ tests/cli/test_model_command.py
@@ -6,7 +6,7 @@ import os
 import pytest
 from click.testing import CliRunner
 
-from rllm.experimental.cli.main import cli
+from rllm.cli.main import cli
 from rllm.experimental.eval.config import RllmConfig, load_config, save_config
